requirements.txt
solana>=0.32.0
solders>=0.20.0
httpx>=0.27.0
websockets>=12.0
python-dotenv>=1.0.0
base58>=2.1.1
pandas>=2.2.0
numpy>=1.26.0
aiofiles>=23.2.1
aiosqlite>=0.20.0
colorlog>=6.8.2
tenacity>=8.3.0